+++ academic-observatory-workflows/academic_observatory_workflows/crossref_fundref_telescope/tasks.py
@@ -232,7 +232,6 @@ def add_dataset_releases(release: Dict, api_bq_dataset_id: str) -> None:
 
     release = CrossrefFundrefRelease.from_dict(release)
     api = DatasetAPI(bq_project_id=release.cloud_workspace.project_id, bq_dataset_id=api_bq_dataset_id)
-    api.seed_db()
     now = pendulum.now()
     dataset_release = DatasetRelease(
         dag_id=release.dag_id,
