--- academic-observatory-workflows/academic_observatory_workflows/data_quality_workflow/tests/fixtures/people20230101.jsonl
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:90eb993257b8b5dfac45998cdb83935edc02743e75d49a4e9ea2c5a56eedbefe
-size 770
