--- academic-observatory-workflows/academic_observatory_workflows/data_quality_workflow/tests/fixtures/people20230108.jsonl
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9065609264312d188c41ca5d0e8d91d3a7448ea97c644dd6cc683c284274ce85
-size 465
