--- academic-observatory-workflows/academic_observatory_workflows/data_quality_workflow/tests/fixtures/people_schema.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e9bfba90202d6aee71fb7c9b31e436404908926c0bde1596c2624ad951bfccc1
-size 305
