--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/authors/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:afc8a5ccf0622898e380acfc448e92c5c58b1dabbc24e89765c34b6509cb8afe
-size 3109
