--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/concepts/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:988200d47a2735a709a2e3f7ecd2a2c23fe67ea2176cf7baea9f48fbd5e84362
-size 56079
