--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/domains/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:785301235a236406cd5e0ebef830ae160e0a3f7530d5525098779c49301607a0
-size 1762
