--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/fields/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ee446db3dac5cb382b2a8583a2c06f608bac789c505c240ffa1430a9724a2f0c
-size 6153
