--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/merged_ids/authors/2023-04-02.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:eb3c189ce09c0aee0796a20d919fb0f481f6aa9e9a6a94499d69646760a80aa8
-size 63
