--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/merged_ids/institutions/2023-04-02.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4076a376b8256e72118175789aa925b19672665101226697ad500a72de9e6e32
-size 59
