--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/merged_ids/sources/2023-04-02.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c635fc862bffd8291b8c36db73f2f503ecf1454b98edd78c80348f29c3b67700
-size 62
