--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/merged_ids/works/2023-04-02.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:feb2305fb8dc360709f0d3a06c4a5df87536c770fdd03e067c69de2858c8aeac
-size 63
