--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/publishers/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c1fef21cbc3fb63bdc8cf8f88b2372bd3286f12fb0db2c3739048043ae290f34
-size 7843
