--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/subfields/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c0883ea24948a3317eeebaf3c824e0bfc689ceaa14111fcc785b8a9e1f06799a
-size 8825
