--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/data/works/updated_date=2023-04-02/part_000.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:302bd45f5ae92e64a4154208913c7c7198555e29243c5abc82b7237ea2de22c5
-size 61634
