--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/authors.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a2e0b1b2ce5d9551695ad92a414654e40333274efaf8c3bb1fcbb61e151edfbb
-size 4667
