--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/concepts.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e7af41245c3279bdef11dce94afa87282ea22463eed22ad692c2a765c1f5cee5
-size 66011
