--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/domains.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a1a908e3b9b2386ec74ec673bb4ba7370c6a9a50ff92158fa7d61eeab21a2c75
-size 1715
