--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/funders.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:db96ee5d6d69e00dc05417a4a9d00afdce8f30c5def88c6b1a851aa6e46261d0
-size 9495
