--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/institutions.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f425b53afb6e0438c744b88c6bb4ac13d832e5ee6971ed1d27fde85fe3e92aff
-size 44617
