--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/publishers.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:92ec4725723722d79c00b09ad8b8c0c063e328acb202e4a3488b8a813b67be11
-size 10956
