--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/sources.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:930cff6e10eed80bff91c2b580ea225889f2aeb084bf516e77a4a4d7168764b3
-size 26440
