--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/subfields.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1d6d575470258f01dc0427ce30c98167c060c064841f9c87e1df6a349014a988
-size 8780
