--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/topics.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5b1cee733f2342281b6231ac14cd7e43f953c273e37a58f5f70e5b840fc56d2c
-size 18137
