--- academic-observatory-workflows/academic_observatory_workflows/openalex_telescope/tests/fixtures/2023-04-02/expected/works.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:dd72a639690204de8fe75f8521924b79935882eaf7a6338b93c017858e6d4ddd
-size 64451
