+++ academic-observatory-workflows/academic_observatory_workflows/pubmed_telescope/tests/test_telescope.py
@@ -284,7 +284,6 @@ class TestPubMedTelescope(SandboxTestCase):
                 test_run=True,
             )
             api = DatasetAPI(bq_project_id=env.cloud_workspace.project_id, bq_dataset_id=test_params.api_bq_dataset_id)
-            api.seed_db()
 
             main_table_id = bq_table_id(
                 project_id=env.cloud_workspace.project_id,
