+++ academic-observatory-workflows/academic_observatory_workflows/ror_telescope/tests/test_telescope.py
@@ -159,6 +159,5 @@ class TestRorTelescope(SandboxTestCase):
             api = DatasetAPI(
                 bq_project_id=test_params.cloud_workspace.project_id, bq_dataset_id=test_params.api_bq_dataset_id
             )
-            api.seed_db()
             dataset_releases = api.get_dataset_releases(dag_id=test_params.dag_id, entity_id="ror")
             self.assertEqual(len(dataset_releases), 2)
